changement coords de int à float (pourquoi n'y ais-je pas pensé avant) pattern fonctionnels

+++ core/src/com/touhou625/projectile/Bullet.java
@@ -4,11 +4,11 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
 
 public class Bullet extends Projectile {
 
-    public Bullet(int xStart, int yStart, int id, float factor, float angle, double[] direction) {
+    public Bullet(double xStart, double yStart, int id, float factor, float angle, double[] direction) {
         super(xStart, yStart, id, factor, angle, direction);
     }
 
-    public Bullet(int xStart, int yStart, float factor, TextureRegion sprite, float angle, double[] direction) {
+    public Bullet(double xStart, double yStart, float factor, TextureRegion sprite, float angle, double[] direction) {
         super(xStart, yStart, sprite, factor, angle, direction);
     }
